+++ test/samples/object-properties-no-named-function-expressions.js
@@ -1,12 +1,14 @@
 module.exports = [
 	{
 		description: 'transpiles shorthand properties',
+		options: { namedFunctionExpressions: false },
 		input: `obj = { x, y }`,
 		output: `obj = { x: x, y: y }`
 	},
 
 	{
 		description: 'transpiles shorthand methods',
+		options: { namedFunctionExpressions: false },
 
 		input: `
 			obj = {
@@ -15,12 +17,13 @@ module.exports = [
 
 		output: `
 			obj = {
-				foo: function foo () { return 42; }
+				foo: function () { return 42; }
 			}`
 	},
 
 	{
 		description: 'transpiles shorthand methods with quoted names (#82)',
+		options: { namedFunctionExpressions: false },
 
 		input: `
 			obj = {
@@ -29,12 +32,13 @@ module.exports = [
 
 		output: `
 			obj = {
-				'foo-bar': function foo_bar () { return 42; }
+				'foo-bar': function () { return 42; }
 			}`
 	},
 
 	{
 		description: 'transpiles shorthand methods with reserved names (!68)',
+		options: { namedFunctionExpressions: false },
 
 		input: `
 			obj = {
@@ -43,12 +47,13 @@ module.exports = [
 
 		output: `
 			obj = {
-				catch: function catch$1 () { return 42; }
+				catch: function () { return 42; }
 			}`
 	},
 
 	{
 		description: 'transpiles shorthand methods with numeric or string names (#139)',
+		options: { namedFunctionExpressions: false },
 
 		input: `
 			obj = {
@@ -65,34 +70,35 @@ module.exports = [
 
 		output: `
 			obj = {
-				0: function () {},
-				5: function () {},
-				80: function () {},
-				.12e3: function () {},
-				491: function () {},
-				12e34: function () {},
-				0xFFFF: function () {},
-				"a string": function astring() {},
-				"var": function var$1() {},
+				0: function() {},
+				5: function() {},
+				80: function() {},
+				.12e3: function() {},
+				491: function() {},
+				12e34: function() {},
+				0xFFFF: function() {},
+				"a string": function() {},
+				"var": function() {},
 			}`
 	},
 
 	{
 		description: 'shorthand properties can be disabled with `transforms.conciseMethodProperty === false`',
-		options: { transforms: { conciseMethodProperty: false } },
+		options: { namedFunctionExpressions: false, transforms: { conciseMethodProperty: false } },
 		input: `var obj = { x, y, z () {} }`,
 		output: `var obj = { x, y, z () {} }`
 	},
 
 	{
 		description: 'computed properties can be disabled with `transforms.computedProperty === false`',
-		options: { transforms: { computedProperty: false } },
+		options: { namedFunctionExpressions: false, transforms: { computedProperty: false } },
 		input: `var obj = { [x]: 'x' }`,
 		output: `var obj = { [x]: 'x' }`
 	},
 
 	{
 		description: 'transpiles computed properties without spacing (#117)',
+		options: { namedFunctionExpressions: false },
 
 		input: `
 			if (1)
@@ -102,28 +108,11 @@ module.exports = [
 		`,
 		output: `
 			if (1)
-				{ console.log(JSON.stringify(( _obj = {}, _obj['com'+'puted'] = 1, _obj['foo'] = 2, _obj )));
-					var _obj; }
+				{ console.log(JSON.stringify(( obj = {}, obj['com'+'puted'] = 1, obj['foo'] = 2, obj )));
+					var obj; }
 			else
-				{ console.log(JSON.stringify(( _obj$1 = {}, _obj$1['bar'] = 3, _obj$1 )));
-			var _obj$1; }
-		`
-	},
-
-	{
-		description: 'avoids shadowing free variables with method names (#166)',
-
-		input: `
-			let x = {
-				foo() { return foo },
-				bar() {}
-			}
-		`,
-		output: `
-			var x = {
-				foo: function foo$1() { return foo },
-				bar: function bar() {}
-			}
+				{ console.log(JSON.stringify(( obj$1 = {}, obj$1['bar'] = 3, obj$1 )));
+			var obj$1; }
 		`
 	},
 ];